requirements.txt
pyyaml>=6.0
requests>=2.31.0
google-analytics-data>=0.18.0
